Refactoring, including extracting the duplicate code
