+++ packages/template-retail-react-app/package.json
@@ -12,9 +12,9 @@
     "analyze-build": "cross-env MOBIFY_ANALYZE=true npm run build",
     "build": "npm run build-translations && pwa-kit-dev build",
     "build-translations": "npm run extract-default-translations && npm run compile-translations",
-    "compile-translations": "node ./scripts/compile-translations.js translations",
-    "compile-translations:pseudo": "node ./scripts/compile-pseudo-translation.js translations/en-US.json",
-    "extract-default-translations": "node ./scripts/extract-default-messages.js en-US",
+    "compile-translations": "node ./scripts/translations/compile-folder.js translations",
+    "compile-translations:pseudo": "node ./scripts/translations/compile-pseudo.js translations/en-US.json",
+    "extract-default-translations": "node ./scripts/translations/extract-default-messages.js en-US",
     "format": "pwa-kit-dev format \"**/*.{js,jsx}\"",
     "lint": "pwa-kit-dev lint \"**/*.{js,jsx}\"",
     "lint:fix": "npm run lint -- --fix",
