+++ packages/template-retail-react-app/scripts/translations/compile-folder.js
@@ -7,21 +7,13 @@
  */
 /* eslint @typescript-eslint/no-var-requires: "off" */
 const {exec} = require('child_process')
-const fs = require('fs')
-const path = require('path')
-
-const packagePath = path.join(process.cwd(), 'package.json')
-const pkgJSON = JSON.parse(fs.readFileSync(packagePath))
-
-const overridesDir = pkgJSON.ccExtensibility?.overridesDir
-const outputFolder = overridesDir
-    ? path.join(overridesDir, 'app/static/translations/compiled')
-    : 'app/static/translations/compiled'
-
-const inputFolder = process.argv[2]
+const {getOutputFolder} = require('./utils')
 
 const main = () => {
+    const inputFolder = process.argv[2]
+    const outputFolder = getOutputFolder()
     const command = `formatjs compile-folder --ast ${inputFolder} ${outputFolder}`
+
     console.log('Compiling translations into the folder:', outputFolder)
     exec(command, (err) => {
         if (err) {
