+++ packages/template-retail-react-app/scripts/translations/compile-pseudo.js
@@ -7,25 +7,15 @@
  */
 /* eslint @typescript-eslint/no-var-requires: "off" */
 const {exec} = require('child_process')
-const fs = require('fs')
-const path = require('path')
-
-const packagePath = path.join(process.cwd(), 'package.json')
-const pkgJSON = JSON.parse(fs.readFileSync(packagePath))
-
-const overridesDir = pkgJSON.ccExtensibility?.overridesDir
-const outputFolder = overridesDir
-    ? path.join(overridesDir, 'app/static/translations/compiled')
-    : 'app/static/translations/compiled'
-
-const inputFile = process.argv[2]
+const {getOutputFolder} = require('./utils')
 
 const main = () => {
+    const inputFile = process.argv[2]
     const locale = 'en-XA'
-    const outputFile = `${outputFolder}/${locale}.json`
+    const outputFile = `${getOutputFolder()}/${locale}.json`
     const command = `formatjs compile --ast ${inputFile} --out-file ${outputFile} --pseudo-locale ${locale}`
-    console.log('Compiling pseudo translation into the file:', outputFile)
 
+    console.log('Compiling pseudo translation into the file:', outputFile)
     exec(command, (err) => {
         if (err) {
             console.error(err)
